backup/server-v1/pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = "A simple weather application"
authors = ["IgorLandry1997 <njoungangigor@gmail.com>"]
license = "MIT"
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.2.1,<6.0.0"
celery = ">=5.5.2,<6.0.0"
django-prometheus = ">=2.3.1,<3.0.0"
django-celery-results = ">=2.6.0,<3.0.0"
django-celery-beat = ">=2.8.1,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-decouple = "^3.8"
django-cors-headers = "^4.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
